+++ executor-service/src/main/resources/logback.xml
@@ -5,12 +5,12 @@
             <pattern>%d{yyyy-MM-dd HH:mm:ss.SSS} %-5level [%thread] %msg%n</pattern>
         </encoder>
     </appender>
-    <appender name="db" class="executor.service.appender.DataBaseAppender">
-        <connection class="executor.service.appender.manager.LogDatabaseManagerImpl">
+    <appender name="db" class="executor.service.appender.LogStorageAppender">
+        <logStorageManager class="executor.service.appender.manager.LogDatabaseManager">
             <url>jdbc:h2:/tmp/logs</url>
             <username>root</username>
             <password>root</password>
-        </connection>
+        </logStorageManager>
     </appender>
     <root level="INFO">
         <appender-ref ref="stdout"/>
